==> notebooks/11_model_training_chek.ipynb <==
import sys
import os

# Add the parent directory to the Python path
sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "..")))
import src.config as config

import joblib
from hsml.model_schema import ModelSchema
from hsml.schema import Schema
from sklearn.metrics import mean_absolute_error

import src.config as config
from src.data_utils import transform_ts_data_info_features_and_target
from src.inference import (
    fetch_days_data,
    get_hopsworks_project,
    load_metrics_from_registry,
    load_model_from_registry,
)
from src.pipeline_utils import get_pipeline

from datetime import datetime, timedelta, timezone
from zoneinfo import ZoneInfo
import hopsworks
import numpy as np
import pandas as pd
from hsfs.feature_store import FeatureStore

import src.config as config
from src.data_utils import transform_ts_data_info_features

def get_feature_store() -> FeatureStore:
    project = get_hopsworks_project()
    return project.get_feature_store()

def fetch_days_data(days):
    current_date = pd.to_datetime(datetime.now(timezone.utc)).tz_convert("America/New_York")
    fetch_data_from = current_date - timedelta(days=(days))
    fetch_data_to = current_date - timedelta(days=0)
    print(fetch_data_from, fetch_data_to)
    print("ghcgchcgfc")
    fs = get_feature_store()
    fg = fs.get_feature_group(name=config.FEATURE_GROUP_NAME, version=1)

    query = fg.select_all()
    # query = query.filter((fg.pickup_hour >= fetch_data_from))
    df = query.read()
    cond = (df["pickup_hour"] >= fetch_data_from) & (df["pickup_hour"] <= fetch_data_to)
    print(sum(cond))
    return df[cond]
    # return df

print(f"Fetching data from group store ...")
ts_data = fetch_days_data(180)

ts_data
